results: Move annotation_table.jsx file
