+++ app/assets/javascripts/Components/Result/annotation_table.jsx
